requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9
sentry-sdk==2.34.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.47.2
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2

SQLAlchemy~=2.0.42